➕ ban + addemoji
